requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow
apache-airflow-providers-amazon==8.3.1
apache-airflow-providers-cncf-kubernetes==7.3.0
apache-airflow-providers-postgres==5.5.2
astro-provider-venv